:recycle: [#5287] Restructure celery module
Break up the code according to logical domains, which shall make
the refactor to tooling from maykin-common easier to diff.
