+++ bin/celery_beat.sh
@@ -10,8 +10,7 @@ export OTEL_SERVICE_NAME="${OTEL_SERVICE_NAME:-openforms-scheduler}"
 mkdir -p celerybeat
 
 echo "Starting celery beat"
-exec celery --app openforms  --workdir src beat \
+exec celery --app openforms.celery --workdir src beat \
     -l $LOGLEVEL \
     -s ../celerybeat/beat \
     --pidfile=  # empty on purpose, see #1182
-
