+++ src/openforms/tests/test_celery_health_check.py
@@ -8,7 +8,7 @@ from django.test import TestCase
 
 from maykin_common.config import config
 
-from openforms.celery import READINESS_FILE
+from openforms.celery.probes import READINESS_FILE
 
 # real, working Celery broker URL. In CI, the envvar CELERY_BROKER_URL is deliberately
 # set to a broken configuration to prevent broken test isolation, but the tests here
